refactor(game-screens): move game screens out of widgets
Relocate full-screen diplomacy, production, and technology screens into app/lib/features/game/screens and update routes, widgetbook, and tests to use the new screen path so widgets and screens are structurally separated for issue #1878.

Made-with: Cursor
